Major refactor: simplify internal Deck model
Card types are now entirely cosmetic: within a `Deck`, each card is
just a `u8`, which has the nice property of ensuring that the entire
`Deck` struct is only 54 bytes, which is not at all too large to copy.

This change enabled me to get rid of the type parameter on Deck,
which in turn made the actual impls much more straightforward.

I'm not entirely happy about the state of the demo tests, because
it makes me unhappy to have to assume that there was a misprint.
All the same, it really looks like my code is doing the right thing.