BROKEN: implement encryption and decryption
Something very weird is going on with encryption; I don't at all
know why I end up with "KIRAK SFJAZ" instead of "KIRAK SFJAN",
particularly since _decryption_ on that phrase works right!

It's as if the keystream is producing different values in that
position, which really shouldn't happen.

In the meantime, `fn crypt` is a thing of beauty.