BROKEN: attempt to use a self-referential struct for Separated
I still think that this idea has some promise; nothing in the rules
actually prohibits it, AFAIK. However, it's time to give up this
line of approach; after spending several hours banging my head on
this problem, it's just not working. Whatever syntax it requires
is currently beyond me.

I don't actually need a fully-generic iterator combinator, here.
It would be fine to just collect the result of Separated into a
String, and I think that would make a lot of these problems just
kind of go away. That's the tack I'm going to take next, anyway.