get everything building again
A somewhat uglier implementation of the Separate trait was all that
ended up being necessary to get this compiling again. Of course,
it's wasteful in the even that after calling that, we want to do
more character-based processing, but at this point it's not worth
investing the time to implement the pure solution.

Note: I haven't yet attempted to build the tests, let alone run them.