get test compiling again
That isn't to say that they're all working: the encrypt/decrypt
tests in particular are all failing, and demonstrating some kind
of weird drift over time. I'm sure it's a relic of missing an
index base somewhere, expecting 0 and using 1 or vice versa.

Still, with the new tests, I'm fairly confident that the textbyte
package is doing what it's supposed to be doing.