get all the tests passing
Had a hunch about the drift; corrected it; suddenly all the tests
passed. I'm now in a state where I'm not sure why exactly this code
is doing the right thing, but I've got a fairly high confidence that
it in fact is. (I really had a different idea in my head about how
the modulo operations would work.)